implement projective ring wip
