+++ primitive/crypto/src/dress.rs
@@ -1,2 +1,3 @@
 pub mod basic;
+pub mod curve;
 pub mod field;
