+++ primitive/crypto/src/dress/curve.rs
@@ -1,8 +1,11 @@
-mod coordinate;
+mod group;
+mod ring;
 
 #[macro_export]
 macro_rules! curve_operation {
     ($curve:ident, $field:ident, $a:ident, $b:ident, $projective:ident, $affine:ident) => {
+        use zero_crypto::arithmetic::coordinate::projective::*;
+
         impl Curve for $curve {
             type ScalarField = $field;
 
@@ -12,7 +15,7 @@ macro_rules! curve_operation {
         }
 
         impl Projective for $projective {
-            
+
         }
     }
 }
