--- primitive/crypto/src/dress/curve/coordinate.rs
@@ -1,12 +1,0 @@
-#[macro_export]
-macro_rules! coordinate_operation {
-    ($curve:ident) => {
-        pub struct ProjectiveCoordinate<E: $curve> {
-            pub(crate) x: E::ScalarField,
-            pub(crate) y: E::ScalarField,
-            pub(crate) z: E::ScalarField,
-        }
-    }
-}
-
-pub use coordinate_operation;
